ci: 🎡 Fix mac build
