+++ Sources/Hoods/Extensions/Blocks+Extension.swift
@@ -1,0 +1,10 @@
+import Blocks
+
+public extension MailtoComponents {
+    init(recipient: String, subject: String, body: String) {
+        self.init()
+        self.recipient = recipient
+        self.subject = subject
+        self.body = body
+    }
+}
